+++ mycodo/config.py
@@ -325,34 +325,47 @@ MATH_INFO = {
     }
 }
 
+method_dep_base = [
+    ('apt', 'unzip', 'unzip'),
+    ('bash-commands',
+     ['/var/mycodo-root/mycodo/mycodo_flask/static/js/user_js/highcharts-9.1.2.js'],
+     [
+        'wget --no-clobber https://code.highcharts.com/zips/Highcharts-9.1.2.zip',
+        'unzip Highcharts-9.1.2.zip -d Highcharts-9.1.2',
+        'cp Highcharts-9.1.2/code/highcharts.js /var/mycodo-root/mycodo/mycodo_flask/static/js/user_js/highcharts-9.1.2.js',
+        'cp Highcharts-9.1.2/code/highcharts.js.map /var/mycodo-root/mycodo/mycodo_flask/static/js/user_js/highcharts-9.1.2.js.map',
+        'rm -rf Highcharts-9.1.2'
+     ])
+]
+
 # Method info
 METHOD_INFO = {
     'Date': {
         'name': lazy_gettext('Time/Date'),
-        'dependencies_module': []
+        'dependencies_module': method_dep_base
     },
     'Duration': {
         'name': lazy_gettext('Duration'),
-        'dependencies_module': []
+        'dependencies_module': method_dep_base
     },
     'Daily': {
         'name': "{} ({})".format(lazy_gettext('Daily'), lazy_gettext('Time-Based')),
-        'dependencies_module': []
+        'dependencies_module': method_dep_base
     },
     'DailySine': {
         'name': "{} ({})".format(lazy_gettext('Daily'), lazy_gettext('Sine Wave')),
-        'dependencies_module': []
+        'dependencies_module': method_dep_base
     },
     'DailyBezier': {
         'name': "{} ({})".format(lazy_gettext('Daily'), lazy_gettext('Bezier Curve')),
         'dependencies_module': [
             ('apt', 'libatlas-base-dev', 'libatlas-base-dev'),
             ('apt', 'python3-numpy', 'python3-numpy')
-        ]
+        ] + method_dep_base
     },
     'Cascade': {
         'name': lazy_gettext('Method Cascade'),
-        'dependencies_module': []
+        'dependencies_module': method_dep_base
     }
 }